+++ src/main/java/chess/controller/GameController.java
@@ -1,9 +1,9 @@
 package chess.controller;
 
-import chess.domain.ChessBoard;
-import chess.domain.ChessGame;
-import chess.domain.CommandType;
-import chess.domain.GameCommand;
+import chess.domain.chessboard.ChessBoard;
+import chess.domain.game.ChessGame;
+import chess.domain.command.CommandType;
+import chess.domain.command.GameCommand;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.generator.NormalPiecesGenerator;
