+++ src/main/java/chess/domain/command/CommandType.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import java.util.Arrays;
