+++ src/main/java/chess/domain/command/GameCommand.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.command;
 
 import chess.domain.position.Position;
 import java.util.ArrayList;
