+++ src/main/java/chess/domain/state/Ready.java
@@ -1,8 +1,8 @@
 package chess.domain.state;
 
-import chess.domain.ChessBoard;
-import chess.domain.CommandType;
-import chess.domain.GameCommand;
+import chess.domain.chessboard.ChessBoard;
+import chess.domain.command.CommandType;
+import chess.domain.command.GameCommand;
 
 public final class Ready implements State {
