+++ src/main/java/chess/domain/state/Running.java
@@ -1,8 +1,8 @@
 package chess.domain.state;
 
-import chess.domain.ChessBoard;
-import chess.domain.CommandType;
-import chess.domain.GameCommand;
+import chess.domain.chessboard.ChessBoard;
+import chess.domain.command.CommandType;
+import chess.domain.command.GameCommand;
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
