+++ src/test/java/chess/domain/game/ScoreCalculatorTest.java
@@ -1,7 +1,9 @@
-package chess.domain;
+package chess.domain.game;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import chess.domain.chessboard.ChessBoard;
+import chess.domain.game.ScoreCalculator;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.King;
