+++ src/test/java/chess/domain/piece/PiecesTest.java
@@ -1,10 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Rook;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
